Revert "- Collect SSH key functions in one file"
This reverts commit 5153e27d6c9126d1f64db33130788ff0246de9bb.
